# apps/api/openapi.yaml
openapi: 3.0.1
info:
  title: In Midst My Life API
  description: |
    Dynamic interactive CV/résumé system with mask-based identity filtering and narrative generation.

    ## Key Features
    - **Profile Management**: CRUD operations for user profiles with temporal epochs and career stages
    - **Mask System**: Context-specific identity masks with cognitive/expressive/operational ontologies
    - **CV Components**: Experiences, educations, skills, projects, credentials, and attestations
    - **Narrative Generation**: AI-powered résumé narrative blocks with weighting and LLM ranking
    - **Export Formats**: JSON-LD (semantic web), PDF (résumé), Verifiable Credentials (W3C standard)
    - **GraphQL API**: Unified query interface with subscriptions support
    - **Authentication**: JWT bearer tokens with role-based access control
    - **Caching**: Redis-backed (with in-memory fallback) for taxonomy endpoints

    ## Authentication

    Most endpoints require JWT bearer token authentication. Include in request:
    ```
    Authorization: Bearer <token>
    ```

    Tokens include permission claims for fine-grained access control.

  version: 0.2.0
  contact:
    name: API Support
    url: https://github.com/anthropics/in-midst-my-life
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.inmidstmylife.com
    description: Production server
...
paths:
...
  /profiles:
    post:
      tags: [Profiles]
      summary: Create new profile
      operationId: createProfile
      security:
        - bearerAuth: [WRITE_PROFILE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Profiles]
      summary: List all profiles
      operationId: listProfiles
      security:
        - bearerAuth: [READ_PROFILE]
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
      responses:
        '200':
          description: Profiles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /profiles/validate:
    post:
      tags: [Profiles]
      summary: Validate profile data
      description: Validate profile structure without persisting
      operationId: validateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          description: Profile is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  valid:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
  /profiles/{id}:
    get:
      tags: [Profiles]
      summary: Get profile by ID
      operationId: getProfile
      security:
        - bearerAuth: [READ_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Profiles]
      summary: Update profile
      operationId: updateProfile
      security:
        - bearerAuth: [WRITE_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Profiles]
      summary: Delete profile
      operationId: deleteProfile
      security:
        - bearerAuth: [DELETE_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Profile deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /profiles/{id}/masks/select:
    post:
      tags: [Profiles]
      summary: Select identity mask for context
      operationId: selectMask
      security:
        - bearerAuth: [READ_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contexts, tags]
              properties:
                contexts:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Best masks selected for context
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  selectedMasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mask'
  /profiles/{id}/narrative:
    post:
      tags: [Profiles]
      summary: Generate narrative blocks
      description: Generate AI-powered résumé narrative blocks with optional mask filtering
      operationId: generateNarrative
      security:
        - bearerAuth: [READ_NARRATIVE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contexts, tags]
              properties:
                maskId:
                  type: string
                contexts:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 50
                includeMeta:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Narrative blocks generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/NarrativeBlock'
                  meta:
                    $ref: '#/components/schemas/NarrativeMeta'
  /profiles/{id}/narratives:
    get:
      tags: [Profiles]
      summary: Get narrative history
      operationId: getNarratives
      security:
        - bearerAuth: [READ_NARRATIVE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, approved, rejected]
      responses:
        '200':
          description: Narratives retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NarrativeSnapshot'
  /profiles/{id}/narratives/approved:
    get:
      tags: [Profiles]
      summary: Get approved narratives
      operationId: getApprovedNarratives
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approved narratives retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NarrativeSnapshot'
  /profiles/{id}/narratives/{narrativeId}:
    patch:
      tags: [Profiles]
      summary: Update narrative
      operationId: updateNarrative
      security:
        - bearerAuth: [WRITE_NARRATIVE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: narrativeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blocks:
                  type: array
                  items:
                    $ref: '#/components/schemas/NarrativeBlock'
      responses:
        '200':
          description: Narrative updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativeSnapshot'
  /profiles/{id}/narratives/{narrativeId}/approve:
    post:
      tags: [Profiles]
      summary: Approve narrative
      operationId: approveNarrative
      security:
        - bearerAuth: [APPROVE_NARRATIVE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: narrativeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Narrative approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativeSnapshot'
  /profiles/{id}/agent-tokens:
    post:
      tags: [Profiles]
      summary: Create agent token
      operationId: createAgentToken
      security:
        - bearerAuth: [WRITE_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                expiresIn:
                  type: integer
      responses:
        '201':
          description: Agent token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  token:
                    type: string

    get:
      tags: [Profiles]
      summary: List agent tokens
      operationId: listAgentTokens
      security:
        - bearerAuth: [READ_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent tokens retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  tokens:
                    type: array
                    items:
                      type: object
  /profiles/{id}/agent-tokens/{tokenId}:
    delete:
      tags: [Profiles]
      summary: Revoke agent token
      operationId: deleteAgentToken
      security:
        - bearerAuth: [WRITE_PROFILE]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  # ============ CV COMPONENT ENDPOINTS ============
...
components:
  schemas:
    Profile:
      type: object
      required: [identityId, slug]
      properties:
        id:
          type: string
          format: uuid
        identityId:
          type: string
          format: uuid
        slug:
          type: string
        displayName:
          type: string
        title:
          type: string
        headline:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProfileResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          $ref: '#/components/schemas/Profile'
    ProfileListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    Mask:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ontology:
          type: string
          enum: [cognitive, expressive, operational]
        functional_scope:
          type: string
        stylistic_parameters:
          type: object
        activation_rules:
          type: object
        filters:
          type: object
        redaction:
          type: object
...
    NarrativeBlock:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        weight:
          type: number
        templateId:
          type: string
    NarrativeSnapshot:
      type: object
      properties:
        id:
          type: string
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/NarrativeBlock'
        status:
          type: string
          enum: [draft, approved, rejected]
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        metadata:
          type: object
    NarrativeMeta:
      type: object
      properties:
        maskId:
          type: string
        mask:
          type: object
        contexts:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        timelineCount:
          type: integer
        generatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token with permission claims

    agentAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AI agent token for autonomous operations
security:
  - bearerAuth: []